migrate things to use local imports and `__init__.py`, finally
