+++ main.py
@@ -1,0 +1,4 @@
+from ferris_elf import main
+
+if __name__ == "__main__":
+    main()
